requirements.txt
pytest~=8.2.2
pyyaml~=6.0.1

attr~=0.3.2
requests~=2.32.3
dotmap~=1.3.30